chore: Move revision registries to a new package
